Cleaning up migration command. Making small changes to models.
