+++ admin/nodes/views.py
@@ -163,8 +163,8 @@ class NodeDeleteView(PermissionRequiredMixin, NodeDeleteBase):
                 osf_flag = NodeLog.NODE_CREATED
             elif not node.is_registration:
                 node.is_deleted = True
-                node.deleted_date = timezone.now()
                 node.deleted = timezone.now()
+                node.deleted_date = node.deleted
                 flag = NODE_REMOVED
                 message = 'Node {} removed.'.format(node.pk)
                 osf_flag = NodeLog.NODE_REMOVED
