+++ osf/management/commands/migrate_deleted_date.py
@@ -4,21 +4,29 @@ import logging
 from django.core.management.base import BaseCommand
 from django.db import connection, transaction
 from framework.celery_tasks import app as celery_app
+from framework import sentry
 
 logger = logging.getLogger(__name__)
 
 TABLES_TO_POPULATE_WITH_MODIFIED = [
     'osf_comment',
-    'osf_institution'
+    'osf_institution',
+    'osf_privatelink'
 ]
 
 POPULATE_COLUMNS = [
-    'SET statement_timeout = 10000; UPDATE osf_basefilenode SET deleted = deleted_on WHERE id IN (SELECT id FROM osf_basefilenode WHERE deleted_on IS NOT NULL AND deleted IS NULL LIMIT %s) RETURNING id;',
-    'SET statement_timeout = 10000; UPDATE osf_abstractnode SET deleted= CASE WHEN deleted_date IS NOT NULL THEN deleted_date ELSE last_logged END WHERE id IN (SELECT id FROM osf_abstractnode WHERE is_deleted AND deleted IS NULL LIMIT %s) RETURNING id;',
-    'SET statement_timeout = 10000; UPDATE osf_privatelink PL set deleted = NL.date from osf_nodelog NL, osf_privatelink_nodes pl_n WHERE NL.node_id=pl_n.abstractnode_id AND pl_n.privatelink_id = pl.id and PL.id in (SELECT id FROM osf_privatelink WHERE is_deleted AND deleted IS NULL LIMIT %s) RETURNING PL.id;',
+    """UPDATE osf_basefilenode SET deleted = deleted_on
+    WHERE id IN (SELECT id FROM osf_basefilenode WHERE deleted_on IS NOT NULL AND deleted IS NULL LIMIT %s)
+    RETURNING id;""",
+    """UPDATE osf_abstractnode SET deleted = CASE WHEN deleted_date IS NOT NULL THEN deleted_date ELSE last_logged END
+    WHERE id IN (SELECT id FROM osf_abstractnode WHERE is_deleted AND deleted IS NULL LIMIT %s)
+    RETURNING id;"""
 ]
 
-UPDATE_DELETED_WITH_MODIFIED = 'SET statement_timeout = 10000; UPDATE {} SET deleted=modified WHERE id IN (SELECT id FROM {} WHERE is_deleted AND deleted IS NULL LIMIT {}) RETURNING id;'
+UPDATE_DELETED_WITH_MODIFIED = """UPDATE {} SET deleted=modified
+WHERE id IN (SELECT id FROM {} WHERE is_deleted AND deleted IS NULL LIMIT {}) RETURNING id;"""
+
+CHECK_POPULATED = """SELECT deleted, is_deleted FROM {} WHERE deleted IS NULL AND is_deleted ;"""
 
 
 @celery_app.task(name='management.commands.migrate_deleted_date')
@@ -32,21 +40,30 @@ def populate_deleted(dry_run=False, page_size=1000):
             raise RuntimeError('Dry Run -- Transaction rolled back')
 
 def run_statements(statement, page_size, table):
-    logger.info('Populating deleted column in table {}'.format(table))
+    sentry.log_message('Populating deleted column in table {}'.format(table))
     with connection.cursor() as cursor:
+        cursor.execute(CHECK_POPULATED.format(table), [page_size])
+        rows = cursor.fetchall()
+        if not rows:
+            return
         cursor.execute(statement.format(table, table, page_size))
         rows = cursor.fetchall()
         if not rows:
-            logger.info('Sentry notification that {} is populated'.format(table))
+            sentry.log_message('Deleted field in {} table is populated'.format(table))
 
 def run_sql(statement, page_size):
     table = statement.split(' ')[5]
-    logger.info('Populating deleted column in table {}'.format(table))
+    sentry.log_message('Populating deleted column in table {}'.format(table))
+
     with connection.cursor() as cursor:
+        cursor.execute(CHECK_POPULATED.format(table), [page_size])
+        rows = cursor.fetchall()
+        if not rows:
+            return
         cursor.execute(statement, [page_size])
         rows = cursor.fetchall()
         if not rows:
-            logger.info('Sentry notification that {} is populated'.format(table))
+            sentry.log_message('Deleted field in {} table is populated'.format(table))
 
 class Command(BaseCommand):
     help = '''Populates new deleted field for various models. Ensure you have run migrations
@@ -62,23 +79,23 @@ class Command(BaseCommand):
         parser.add_argument(
             '--page_size',
             type=int,
-            default=1000,
+            default=5000,
             help='How many rows to process at a time',
         )
 
     def handle(self, *args, **options):
         script_start_time = datetime.datetime.now()
-        logger.info('Script started time: {}'.format(script_start_time))
+        sentry.log_message('Script started time: {}'.format(script_start_time))
         logger.debug(options)
 
         dry_run = options['dry_run']
         page_size = options['page_size']
 
         if dry_run:
-            logger.info('DRY RUN')
+            sentry.log_message('DRY RUN')
 
         populate_deleted(dry_run, page_size)
 
         script_finish_time = datetime.datetime.now()
-        logger.info('Script finished time: {}'.format(script_finish_time))
-        logger.info('Run time {}'.format(script_finish_time - script_start_time))
+        sentry.log_message('Script finished time: {}'.format(script_finish_time))
+        sentry.log_message('Run time {}'.format(script_finish_time - script_start_time))
