+++ osf_tests/test_management_commands.py
@@ -10,19 +10,16 @@ from addons.osfstorage import settings as osfstorage_settings
 from api_tests.utils import create_test_file
 from framework.auth import Auth
 from osf.models import QuickFilesNode, RegistrationSchema
-from addons.osfstorage.models import OsfStorageFolder
 from osf_tests.factories import (
     PreprintFactory,
     ProjectFactory,
     RegionFactory,
     UserFactory,
-    CommentFactory,
 )
 from tests.base import DbTestCase
 from osf.management.commands.data_storage_usage import (
     process_usages,
 )
-from osf.management.commands import migrate_deleted_date
 
 
 # Using powers of two so that any combination of file sizes will give a unique total
@@ -33,69 +30,6 @@ def next_file_size():
     while True:
         yield size
         size *= 2
-
-class TestMigrateDeletedDate(DbTestCase):
-
-    def setUp(self):
-        super(TestMigrateDeletedDate, self).setUp()
-        self.region_us = RegionFactory(_id='US', name='United States')
-
-    @pytest.fixture()
-    def project(self, user, is_public=True, is_deleted=False, region=None, parent=None):
-        if region is None:
-            region = self.region_us
-        project = ProjectFactory(creator=user, is_public=is_public, is_deleted=is_deleted)
-        addon = project.get_addon('osfstorage')
-        addon.region = region
-        addon.save()
-
-        return project
-
-    @pytest.fixture()
-    def user(self):
-        return UserFactory()
-
-    def test_populate_with_modified(self):
-        statement = migrate_deleted_date.UPDATE_DELETED_WITH_MODIFIED
-        table = 'osf_comment'
-        user = UserFactory()
-        project = ProjectFactory(creator=user)
-        comment = CommentFactory(user=user, node=project)
-
-        comment.delete(Auth(user), save=True)
-        comment.reload()
-        assert(comment.deleted)
-
-        comment.deleted = None
-        comment.save()
-        migrate_deleted_date.run_statements(statement, 1000, table)
-        comment.reload()
-        assert(comment.deleted)
-        assert(comment.deleted == comment.modified)
-
-    def test_populate_columns(self):
-        statement = migrate_deleted_date.POPULATE_COLUMNS[0]
-        user = UserFactory()
-        project = self.project(user)
-        osf_folder = OsfStorageFolder.objects.filter(target_object_id=project.id)[0]
-
-        project.remove_node(Auth(user))
-        osf_folder.delete()
-        osf_folder.reload()
-        project.reload()
-        assert(osf_folder.deleted)
-        assert(project.deleted)
-
-        project.deleted = None
-        osf_folder.deleted = None
-
-        osf_folder.save()
-        project.save()
-        migrate_deleted_date.run_sql(statement, 1000)
-
-        osf_folder.reload()
-        assert(osf_folder.deleted)
-
 
 class TestDataStorageUsage(DbTestCase):
